expandmatrix reshapes a matrix of values into a 3 column data.frame

expandmatrix takes an oblong matrix of values and expands it into a three column data.frame of row, column, value. This is then easier to plot as a scattergram or is used within categoryplot. It expects to have the year values in the columns = xvalues

Usage

expandmatrix(x)

Arguments

x

a matrix of values

Value

a 3-column matrix of (rows x cols) rows from the input matrix

Examples

x <- matrix(rnorm(25,5,1),nrow=5,ncol=5,dimnames=list(1:5,1:5))
res <- expandmatrix(x)
res
#>     rows cols    value
#> 1_1    1    1 3.055387
#> 1_2    1    2 5.026989
#> 1_3    1    3 5.133698
#> 1_4    1    4 3.218920
#> 1_5    1    5 3.243409
#> 2_1    2    1 6.017416
#> 2_2    2    2 5.415466
#> 2_3    2    3 5.277223
#> 2_4    2    4 4.026511
#> 2_5    2    5 5.478653
#> 3_1    3    1 5.398259
#> 3_2    3    2 4.504542
#> 3_3    3    3 4.836473
#> 3_4    3    4 7.040670
#> 3_5    3    5 5.266679
#> 4_1    4    1 4.560681
#> 4_2    4    2 5.677932
#> 4_3    4    3 5.180344
#> 4_4    4    4 5.569414
#> 4_5    4    5 4.270722
#> 5_1    5    1 3.968250
#> 5_2    5    2 2.824747
#> 5_3    5    3 4.072060
#> 5_4    5    4 4.063095
#> 5_5    5    5 3.115200