parset alters the current base graphics par settings

parset alters the current base graphics par settings to suit a single standard plot. It is merely here to simplify and speed the coding for exploratory base graphics. The font and its size default to 0.85 and font 7 (Times bold). The default values can be seen by typing parset with no brackets in the console. If a different set of par values are needed then the function parsyn() can be used to act as a prompt for the correct syntax. The output to the console can be copied to your script and modified to suit.

Usage

parset(
  plots = c(1, 1),
  cex = 0.75,
  font = 7,
  outmargin = c(0, 0, 0, 0),
  margin = c(0.45, 0.45, 0.05, 0.05),
  byrow = TRUE,
  ...
)

Arguments

plots

vector of number of rows and columns, defaults to c(1,1)

cex

the size of the font used, defaults to 0.85

font

the font used, defaults to 7 which is Times Bold, 6 is Times, 1 is Sans and 2 is Sans Bold.

outmargin

default=c(0,0,0,0) and defines the outer margin used by mtext

margin

default=c(0.45,0.45,0.05,0.05), which avoids whitespace but leaves plenty of room for titles

byrow

should plots be made by row (mfrow; byrow=TRUE, the default), of by column (mfcol; byrow=FALSE)

...

the generic ellipsis allowing for the includion of other graphics arguments such as xaxs="n", etc.

Value

nothing but it changes the base graphics par settings

Examples

if (FALSE) {
parset()
parsyn()
}